% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_resourcePath}
\alias{add_resourcePath}
\title{Add resource paths}
\usage{
add_resourcePath()
}
\value{
NA
}
\description{
Add resource paths
}
\examples{
add_resourcePath()

}
\author{
Samuel Wieczorek
}
