% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_funcs.R
\name{addDatasets}
\alias{addDatasets}
\title{Adds a dataset to the list}
\usage{
addDatasets(object, dataset, name)
}
\arguments{
\item{object}{An instance of the class \code{QFeatures}
Must get TRUE to inherits(object, 'QFeatures')}

\item{dataset}{An instance of class \code{SummarizedExperiment}}

\item{name}{the name to associate to the dataset in the final object}
}
\description{
This function appends a dataset in the list with customization
if necessary
}
