% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_funcs.R
\name{aggregationPept}
\alias{aggregationPept}
\title{Peptide aggregation}
\usage{
aggregationPept(
  data,
  history = NULL,
  sharePept,
  operator,
  considerPept,
  ponderation,
  n,
  aggCol,
  maxIter,
  rmEmptyLines
)
}
\arguments{
\item{data}{A \code{QFeatures}}

\item{history}{The history}

\item{sharePept}{How shared peptides are handled. Either \code{Yes_As_Specific},
\code{Yes_Iterative_Redistribution}, \code{Yes_Simple_Redistribution} or \code{No}}

\item{operator}{A function used for quantitative feature aggregation.
Available functions are \code{Sum}, \code{Mean}, \code{Median}, \code{medianPolish} or
\code{robustSummary}}

\item{considerPept}{A \code{character(1)} defining what peptide to consider.
Available values are \code{allPeptides} and \code{topN}}

\item{ponderation}{A \code{character(1)} defining what to consider to create the
coefficient for redistribution of shared peptides. Available values are
\code{Global}, \code{Condition} or \code{Sample}}

\item{n}{If \code{topN}, specifies the number of peptides to use for each protein}

\item{aggCol}{A \code{character()} of column names from rowdata to be aggregated}

\item{maxIter}{A \code{numeric(1)} setting the maximum number of iteration}

\item{rmEmptyLines}{A \code{logical(1)}, whether empty lines should be
automatically removed after aggregation}
}
\value{
A \code{list} containing the aggregated dataset and the history
}
\description{
Perform aggregation for peptides
}
\examples{
NULL

}
