% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_funcs.R
\name{aggregStatTables}
\alias{aggregStatTables}
\title{Aggregation stats tables}
\usage{
aggregStatTables(data)
}
\arguments{
\item{data}{A \code{QFeatures}}
}
\value{
A \code{list} containing the 2 \code{data.frame}
}
\description{
Make table for aggregation stats
}
\examples{
data(subR25pept, package = "DaparToolshed")
aggregStatTables(subR25pept)

}
