% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_buildDesign.R
\name{build-design}
\alias{build-design}
\alias{mod_buildDesign_ui}
\alias{mod_buildDesign_server}
\alias{mod_buildDesign}
\title{xxxx}
\usage{
mod_buildDesign_ui(id)

mod_buildDesign_server(
  id,
  quantCols,
  remoteReset = reactive({
     FALSE
 }),
  is.enabled = reactive({
     TRUE
 })
)

mod_buildDesign(quantCols)
}
\arguments{
\item{id}{xxx}

\item{quantCols}{xxx}
}
\value{
NA
}
\description{
xxxx
}
