% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Build_enriched_qdata}
\alias{Build_enriched_qdata}
\title{xxxx}
\usage{
Build_enriched_qdata(obj.se, digits = NULL)
}
\arguments{
\item{obj.se}{xx}

\item{digits}{xxx}
}
\value{
NA
}
\description{
xxxx
}
\examples{
NULL

}
