% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDA_funcs.R
\name{build_pairwise_comparison_workbook}
\alias{build_pairwise_comparison_workbook}
\title{Create downloadable table}
\usage{
build_pairwise_comparison_workbook(pval_table, comparison)
}
\arguments{
\item{pval_table}{A \code{data.frame} containing the p-values, adjusted p-values
and logFC}

\item{comparison}{A \code{character(1)}, the name of the comparison}
}
\value{
A workbook to download the table
}
\description{
Create downloadable table as workbook
}
\examples{
NULL

}
