% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDA_funcs.R
\name{build_pval_table}
\alias{build_pval_table}
\title{Build p-value table}
\usage{
build_pval_table(
  data,
  comparison,
  thlogfc,
  thpval,
  calibration_method,
  tooltip_info
)
}
\arguments{
\item{data}{A \code{SummarizedExperiment}}

\item{comparison}{A \code{character(1)}, the name of the comparison}

\item{thlogfc}{A \code{numeric(1)}, the logFC threshold}

\item{thpval}{A \code{numeric(1)}, the p-value threshold}

\item{calibration_method}{A \code{character(1)}, the calibration method}

\item{tooltip_info}{A \code{charcter()}, all selected tooltips}
}
\value{
A \code{data.frame} containing the p-values and important informations
}
\description{
Build the table containing the p-values and important
informations
}
\examples{
NULL

}
