% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_funcs.R
\name{cellmetadataFiltering}
\alias{cellmetadataFiltering}
\title{Cell metadata filtering}
\usage{
cellmetadataFiltering(data, filters, query, history, original_length)
}
\arguments{
\item{data}{A \code{QFeatures}}

\item{filters}{The filter}

\item{query}{A \code{character(1)}, the query for the filtering}

\item{history}{A \code{data.frame}, the history}

\item{original_length}{A \code{numeric(1)}, the length of the dataset prefiltering}
}
\value{
A \code{list} containing the filtered dataset, important values
and the history
}
\description{
Perform cell metadata filtering
}
\examples{
NULL

}
