% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDA_funcs.R
\name{checkLimma}
\alias{checkLimma}
\title{Check conditions for Limma}
\usage{
checkLimma(data)
}
\arguments{
\item{data}{A \code{QFeatures}}
}
\value{
A \code{logical(1)} if limma can be applied
}
\description{
Check if limmae can be applied to a dataset
}
\examples{
data(subR25prot, package = "DaparToolshed")
checkLimma(subR25prot)

}
