% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkNA}
\alias{checkNA}
\title{Check missing values}
\usage{
checkNA(data, i = NULL)
}
\arguments{
\item{data}{A \code{QFeatures}}

\item{i}{The index or name of the assay to check}
}
\value{
A \code{logical(1)}
}
\description{
Check whether there is missing values in a SE assay
}
\examples{
data(subR25prot, package = "DaparToolshed")
checkNA(subR25prot)

}
