% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDA_funcs.R
\name{compute_fdr}
\alias{compute_fdr}
\title{Compute FDR}
\usage{
compute_fdr(pval_table, thpval, thlogfc)
}
\arguments{
\item{pval_table}{A \code{data.frame} containing the p-values, adjusted p-values
and logFC}

\item{thpval}{A \code{numeric(1)}, the p-value threshold}

\item{thlogfc}{A \code{numeric(1)}, the logFC threshold}
}
\value{
A \code{data.frame} containing the p-values and important informations
}
\description{
Compute the FDR
}
\examples{
NULL

}
