% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDA_funcs.R
\name{count_selected_prot}
\alias{count_selected_prot}
\title{Count selected proteins}
\usage{
count_selected_prot(pval_table, thpval, thlogfc)
}
\arguments{
\item{pval_table}{A \code{data.frame} containing the p-values, adjusted p-values
and logFC}

\item{thpval}{A \code{numeric(1)}, the p-value threshold}

\item{thlogfc}{A \code{numeric(1)}, the logFC threshold}
}
\value{
A \code{numeric(1)} containing the number of selected proteins
}
\description{
Count selected proteins
}
\examples{
NULL

}
