% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDA_funcs.R
\name{count_significant}
\alias{count_significant}
\title{Count significant proteins}
\usage{
count_significant(pval_table, comparison)
}
\arguments{
\item{pval_table}{A \code{data.frame} containing the p-values, adjusted p-values
and logFC}

\item{comparison}{A \code{character(1)}, the name of the comparison}
}
\value{
A \code{numeric(1)} containing the number of significant proteins
}
\description{
Count the number of significant proteins
}
\examples{
NULL

}
