% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{countPattern}
\alias{countPattern}
\title{Count pattern}
\usage{
countPattern(dataSE, pattern, level = NULL)
}
\arguments{
\item{pattern}{A \code{character(1)}, pattern to count}

\item{level}{Dataset level}

\item{data}{A \code{SummarizedExperiment}}
}
\value{
A \code{numeric(1)}
}
\description{
Count the number of occurence of a selected metacell tag
}
\examples{
data(subR25prot, package = "DaparToolshed")
countPattern(subR25prot[[length(subR25prot)]],
               pattern = "Missing POV")

}
