% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_funcs.R
\name{dataset-processing}
\alias{dataset-processing}
\alias{keepDatasets}
\alias{keepDatasets,list-method}
\alias{addDatasets}
\alias{addDatasets,list-method}
\title{Datasets processing}
\usage{
addDatasets(object, dataset, name)

keepDatasets(object, range = seq(length(object)))
}
\arguments{
\item{object}{An object of class \code{QFeatures}.}

\item{dataset}{A \code{SummarizedExperiment} object containing the dataset to be
added.}

\item{name}{A \code{character(1)} specifying the name of the new assay.}

\item{range}{A \code{numeric} vector containing the indices of assays to retain.}
}
\value{
A processed \code{QFeatures} object.
}
\description{
This manual page describes methods for manipulating datasets stored in
\link[QFeatures:QFeatures]{QFeatures} objects. A \code{QFeatures} object contains a collection of assays,
each represented as a \code{SummarizedExperiment} object.

The following functions are currently available:

\describe{
\item{\code{addDatasets(object, dataset, name)}}{Adds a new dataset (assay) to a
\code{QFeatures} object and assigns it a user-defined name.}
\item{\code{keepDatasets(object, range)}}{Keeps only selected datasets (assays)
from a \code{QFeatures} object and removes all others.}
}
}
\examples{
NULL

}
