% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrationPlot.R
\name{diffAnaComputeAdjustedPValues}
\alias{diffAnaComputeAdjustedPValues}
\title{Computes the adjusted p-values}
\usage{
diffAnaComputeAdjustedPValues(pval, pi0Method = 1)
}
\arguments{
\item{pval}{The result (p-values) of the differential analysis processed
by limmaCompleteTest}

\item{pi0Method}{The parameter pi0.method of the method adjust.p in the
package cp4p}
}
\value{
The computed adjusted p-values
}
\description{
This function is a wrapper to the function adjust.p from the \code{cp4p} package.
It returns the FDR corresponding to the p-values of the differential
analysis. The FDR is computed with the function p.adjust
}
\examples{
if (interactive()){
library(DaparToolshed)
library(SummarizedExperiment)
data(Exp1_R25_prot, package = "DaparToolshedData")
obj <- Exp1_R25_prot
# Simulate imputation of missing values
obj <- NAIsZero(obj, 1)
obj <- NAIsZero(obj, 2)
qData  <- as.matrix(SummarizedExperiment::assay(obj[[2]]))
sTab <- MultiAssayExperiment::colData(obj)
limma <- limmaCompleteTest(qData, sTab)
df <- data.frame(id = rownames(limma$logFC), logFC = limma$logFC[, 1], pval = limma$P_Value[, 1])

diffAnaComputeAdjustedPValues(pval = limma$P_Value[, 1])
}

}
\author{
Samuel Wieczorek
}
