% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download_dataset.R
\name{download_dataset}
\alias{download_dataset}
\alias{download_dataset_ui}
\alias{download_dataset_server}
\title{dl}
\usage{
download_dataset_ui(id)

download_dataset_server(
  id,
  dataIn = reactive({
     NULL
 }),
  extension = c("xlsx", "qf"),
  filename = "myDataset",
  excel.style = NULL,
  remoteReset = reactive({
     0
 }),
  is.enabled = reactive({
     TRUE
 })
)

download_dataset(dataIn = NULL, filename = "myDataset")
}
\arguments{
\item{id}{internal}

\item{dataIn}{internal}

\item{extension}{Available values are \code{csv} (default), \code{qf} and \code{Excel}.}

\item{filename}{internal}

\item{excel.style}{xxx}

\item{remoteReset}{A \code{logical(1)} which acts as a remote command to reset
the module to its default values. Default is FALSE.}

\item{is.enabled}{xxx}
}
\value{
NA
}
\description{
A shiny Module.
}
\examples{
if (interactive()){
data(Exp1_R25_prot, package = "DaparToolshedData")
shiny::runApp(download_dataset(Exp1_R25_prot))

shiny::runApp(download_dataset())
}

}
