% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Extract_Value}
\alias{Extract_Value}
\title{Extract value with the correct class}
\usage{
Extract_Value(
  value,
  expected_type = c("numeric", "character", "logical", "factor", "integer")
)
}
\arguments{
\item{value}{Value to extract}

\item{expected_type}{Expected value class}
}
\value{
The value with the correct class,
or NA if the coercion is not possible
}
\description{
Extract value with the correct class
}
\examples{
Extract_Value(1, "character")
Extract_Value("test", "numeric")

}
