% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDA_funcs.R
\name{fdr_warning_text}
\alias{fdr_warning_text}
\title{Warning text FDR}
\usage{
fdr_warning_text(fdr, n_significant)
}
\arguments{
\item{fdr}{A \code{numeric(1)}, the FDR}

\item{n_significant}{A \code{numeric(1)}, the number of selected proteins}
}
\value{
Either \code{NULL}, or a \code{character(1)} issuing a warning
}
\description{
Warning text in case the number of selected proteins is too low
compared to the FDR
}
\examples{
NULL

}
