% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDA_funcs.R
\name{fdrProt}
\alias{fdrProt}
\title{PipelineProtein FDR sub-step}
\usage{
fdrProt(history, thpval, FDR, nbSignif)
}
\arguments{
\item{history}{The history}

\item{thpval}{A \code{numeric(1)}, the pval threshold}

\item{FDR}{A \code{numeric(1)}, the FDR}

\item{nbSignif}{A \code{numeric(1)}, the number of significant proteins}
}
\value{
A list containing the history
}
\description{
Do what has to be done at the end of the FDR sub-step of the
PipelineProtein
}
\examples{
NULL

}
