% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDA_funcs.R
\name{get_calibration_method}
\alias{get_calibration_method}
\title{Get calibration method}
\usage{
get_calibration_method(calibration_method, numeric_value = NULL)
}
\arguments{
\item{calibration_method}{A \code{character(1)}, the calibration method}

\item{numeric_value}{A \code{numeric(1)}, the pi0 value if needed}
}
\value{
A \code{character(1)} or a \code{numeric(1)} containing the calibration method
}
\description{
Get the calibration method
}
\examples{
get_calibration_method("Benjamini-Hochberg")

}
