% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDA_funcs.R
\name{Get_Pairwisecomparison_Names}
\alias{Get_Pairwisecomparison_Names}
\title{Comparison names}
\usage{
Get_Pairwisecomparison_Names(data)
}
\arguments{
\item{data}{A \code{data.frame} of the tests}
}
\value{
A \code{character()} containing the name of the comparisons
}
\description{
Get name of the comparisons
}
\examples{
NULL

}
