% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDA_funcs.R
\name{GetdatasetToAnalyze}
\alias{GetdatasetToAnalyze}
\title{Get data to analyse}
\usage{
GetdatasetToAnalyze(data, comparison)
}
\arguments{
\item{data}{A \code{QFeatures}}

\item{comparison}{A \code{character(1)}, the name of the comparison}
}
\value{
A \code{SummarizedExperiment}
}
\description{
Get the dataset with only the conditions of interest
}
\examples{
NULL

}
