% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetFiltersScope}
\alias{GetFiltersScope}
\title{Get filters scope}
\usage{
GetFiltersScope()
}
\value{
A \code{list}
}
\description{
Get the list of possible scopes for filters
}
\examples{
GetFiltersScope()

}
