% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetHistory}
\alias{GetHistory}
\title{Get the history of an assay}
\usage{
GetHistory(dataIn, x)
}
\arguments{
\item{dataIn}{An instance of \code{MultiAssayExperiment} class}

\item{x}{The name of a slot in the object}
}
\value{
A \code{data.frame()}
}
\description{
Get the history of an assay
}
\examples{
NULL

}
