% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDA_funcs.R
\name{getlogFC}
\alias{getlogFC}
\title{Get logFC}
\usage{
getlogFC(data, type = "OnevsOne")
}
\arguments{
\item{data}{A \code{QFeatures}}

\item{type}{A \code{character(1)}, type of contrast to get the logFC of.
Either \code{"OnevsOne"} or \code{"OnevsAll"}}
}
\value{
A \code{data.frame} containing the logFC
}
\description{
Get logFC from a dataset
}
\examples{
data(subR25prot, package = "DaparToolshed")
getlogFC(subR25prot)

}
