% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_funcs.R
\name{GetMECimputMet}
\alias{GetMECimputMet}
\title{Protein MEC imputation methods}
\usage{
GetMECimputMet()
}
\value{
A vector of character
}
\description{
Get imputation methods for proteins for missing MEC
}
\examples{
GetMECimputMet()

}
