% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_funcs.R
\name{GetPOVimputMet}
\alias{GetPOVimputMet}
\title{Protein POV imputation methods}
\usage{
GetPOVimputMet()
}
\value{
A vector of character
}
\description{
Get imputation methods for proteins for missing POV
}
\examples{
GetPOVimputMet()

}
