% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDA_funcs.R
\name{getPValueNonPushed}
\alias{getPValueNonPushed}
\title{Clean p-values}
\usage{
getPValueNonPushed(pval)
}
\arguments{
\item{pval}{A vector of p-values}
}
\value{
A vector of p-values
}
\description{
Get p-values that were not pushed
}
\examples{
NULL

}
