% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{globalvariables}
\alias{globalvariables}
\alias{btn_success_color}
\alias{optionsBtnClass}
\alias{redBtnClass}
\alias{PrevNextBtnClass}
\alias{actionBtnClass}
\alias{listBrewerPalettes}
\alias{btnVentsMasg}
\title{Global variables}
\usage{
btn_success_color

optionsBtnClass

redBtnClass

PrevNextBtnClass

actionBtnClass

listBrewerPalettes

btnVentsMasg
}
\value{
NA
}
\description{
Global constants used by the package for user interface
configuration and default values.
}
\examples{
NULL 

}
