% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_history_dataset.R
\name{history_dataset}
\alias{history_dataset}
\alias{history_dataset_ui}
\alias{history_dataset_server}
\title{history_dataset_ui and history_dataset_server}
\usage{
history_dataset_ui(id)

history_dataset_server(
  id,
  dataIn = reactive({
     NULL
 }),
  remoteReset = reactive({
     0
 }),
  is.enabled = reactive({
     TRUE
 })
)

history_dataset(obj)
}
\arguments{
\item{id}{shiny id}

\item{dataIn}{An instance of the class \code{QFeatures}.}
}
\value{
A shiny app
}
\description{
A shiny Module.
}
\examples{
if (interactive()){
data(Exp1_R25_prot, package = "DaparToolshedData")
shiny::runApp(history_dataset(Exp1_R25_prot))
}

}
\keyword{internal}
