% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrationPlot.R
\name{histPValue_HC}
\alias{histPValue_HC}
\title{Plots a histogram ov p-values}
\usage{
histPValue_HC(pval_ll, bins = 80, pi0 = 1)
}
\arguments{
\item{pval_ll}{xxx}

\item{bins}{xxx}

\item{pi0}{xxx}
}
\value{
A plot
}
\description{
Plots a histogram ov p-values
}
\examples{
if (interactive()){
library(DaparToolshed)
library(SummarizedExperiment)
data(Exp1_R25_prot, package = "DaparToolshedData")
obj <- Exp1_R25_prot
# Simulate imputation of missing values
obj <- NAIsZero(obj, 1)
obj <- NAIsZero(obj, 2)
 qData<- as.matrix(SummarizedExperiment::assay(obj[[2]]))
sTab <- MultiAssayExperiment::colData(obj)
limma <- limmaCompleteTest(qData, sTab)
histPValue_HC(limma$P_Value[1])
}

}
\author{
Samuel Wieczorek
}
