% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDA_funcs.R
\name{hypothesisTestProt}
\alias{hypothesisTestProt}
\title{Hypothesis test}
\usage{
hypothesisTestProt(data, method, history = NULL, logFC_thr, design, ttest_type)
}
\arguments{
\item{data}{A \code{QFeatures}}

\item{method}{A \code{character(1)}, hypothesis test type to use. Must choose
between : \code{"Limma"} and \code{"ttests"}}

\item{history}{The history}

\item{logFC_thr}{A \code{numeric(1)}, the logFC threshold value}

\item{design}{A \code{character(1)}, the type of contrast. Must choose between :
\code{"OnevsOne"} and \code{"OnevsAll"}}

\item{ttest_type}{A \code{character(1)}, the type of t-test. Must choose between :
\code{"Student"} and \code{"Welch"}}
}
\value{
A \code{list} containing the hypothesis test result, the history
and additionnal informations
}
\description{
Perform hypothesis test for a dataset
}
\examples{
NULL

}
