% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_funcs.R
\name{imputationPept}
\alias{imputationPept}
\title{Peptide imputation}
\usage{
imputationPept(
  data,
  method,
  history = NULL,
  dataPirat = NULL,
  extension = NULL,
  alpha.factor = NULL,
  nPcs = NULL
)
}
\arguments{
\item{data}{A \code{QFeatures}}

\item{method}{A \code{character(1)}, imputation method to be used. Must choose
between : \code{Pirat}, \code{impSeq} and \code{BPCA}}

\item{history}{The history}

\item{dataPirat}{The dataset in a Pirat compliant format}

\item{extension}{A \code{character(1)}, the extension for Pirat}

\item{alpha.factor}{A \code{numeric(1)}, the alpha factor for Pirat}

\item{nPcs}{A \code{numeric(1)}, the nPcs for BPCA}
}
\value{
A \code{list} containing the imputed dataset and the history
}
\description{
Perform imputation for peptides
}
\examples{
data(subR25pept, package = "DaparToolshed")
imputationPept(subR25pept,
               method = "impSeq")

}
