% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_funcs.R
\name{imputationProtMEC}
\alias{imputationProtMEC}
\title{Protein MEC imputation}
\usage{
imputationProtMEC(
  data,
  method,
  history = NULL,
  quantile = NULL,
  factor = NULL,
  fixVal = NULL
)
}
\arguments{
\item{data}{A \code{QFeatures}}

\item{method}{A \code{character(1)}, imputation method to be used. Must choose
between : \code{detQuantile} and \code{fixedValue}}

\item{history}{The history}

\item{quantile}{A \code{numeric(1)}, quantile used as the reference value.
Must be a float between 0 and 100}

\item{factor}{A \code{numeric(1)}, multiplying factor to apply to the chosen
quantile value}

\item{fixVal}{A \code{numeric(1)}, fixed value used to impute all MEC}
}
\value{
A \code{list} containing the imputed dataset and the history
}
\description{
Perform imputation for proteins for missing MEC
}
\examples{
data(subR25prot, package = "DaparToolshed")
imputationProtMEC(subR25prot,
                  method = "fixedValue",
                  fixVal = 3)

}
