% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_funcs.R
\name{imputationProtPOV}
\alias{imputationProtPOV}
\title{Protein POV imputation}
\usage{
imputationProtPOV(
  data,
  method,
  history = NULL,
  quantile = NULL,
  factor = NULL,
  n = NULL
)
}
\arguments{
\item{data}{A \code{QFeatures}}

\item{method}{A \code{character(1)}, imputation method to be used. Must choose
between : \code{slsa}, \code{detQuantile} and \code{KNN}}

\item{history}{The history}

\item{quantile}{A \code{numeric(1)}, quantile used as the reference value.
Must be a float between 0 and 100}

\item{factor}{A \code{numeric(1)}, multiplying factor to apply to the chosen
quantile value}

\item{n}{A \code{numeric(1)}, number of neighbors to account for}
}
\value{
A \code{list} containing the imputed dataset and the history
}
\description{
Perform imputation for proteins for missing POV
}
\examples{
data(subR25prot, package = "DaparToolshed")
imputationProtPOV(subR25prot,
                  method = "slsa")

}
