% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{InitializeHistory}
\alias{InitializeHistory}
\title{Initialize the history}
\usage{
InitializeHistory()
}
\value{
An empty \code{data.frame} with 4 columns ('Step', 'Substep', 'Parameter' and 'Value')
}
\description{
This function initializes the history.
}
\examples{
InitializeHistory()

}
