% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isContainedIn}
\alias{isContainedIn}
\title{Object contained in another one}
\usage{
isContainedIn(strA, strB)
}
\arguments{
\item{strA}{Object to find}

\item{strB}{Object to check}
}
\value{
A \code{logical(1)}
}
\description{
Check whether an object contained in another one
}
\examples{
isContainedIn("A", c("A", "B"))
isContainedIn("A", c("ABCDE", "FGHIJ"))

}
