% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_funcs.R
\name{keepDatasets}
\alias{keepDatasets}
\title{Get a subset of the object}
\usage{
keepDatasets(object, range = seq(length(object)))
}
\arguments{
\item{object}{An instance of type list. Must get TRUE to inherits(object, 'list')}

\item{range}{xxx}
}
\description{
This function deletes the items not included in the
range parameter
}
