% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDA_funcs.R
\name{makeDTselectedProt}
\alias{makeDTselectedProt}
\title{Make DT selected proteins}
\usage{
makeDTselectedProt(pval_table, view_adj, comparison)
}
\arguments{
\item{pval_table}{A \code{data.frame} containing the p-values, adjusted p-values
and logFC}

\item{view_adj}{A \code{logical(1)}, whether to show the adjusted p-values or not}

\item{comparison}{A \code{character(1)}, the name of the comparison}
}
\value{
A DT::datatable
}
\description{
Make the DT of selected proteins
}
\examples{
NULL

}
