% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{message_console}
\alias{message_console}
\title{Message console}
\usage{
message_console(
  msg,
  level = "INFO",
  info_text = "INFO",
  warning_text = "WARNING",
  error_text = "ERROR",
  color_info = "grey",
  color_warning = "blue",
  color_error = "red",
  color_other = "black"
)
}
\arguments{
\item{msg}{xxx}

\item{level}{xxx}

\item{info_text}{xxx}

\item{warning_text}{xxx}

\item{error_text}{xxx}

\item{color_info}{xxx}

\item{color_warning}{xxx}

\item{color_error}{xxx}

\item{color_other}{xxx}
}
\value{
xxx
}
\description{
Style R logs in a console in shiny
}
\examples{
NULL

}
