% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_funcs.R
\name{missmechPiratPlot}
\alias{missmechPiratPlot}
\title{Pirat missingness mechanism plot}
\usage{
missmechPiratPlot(data)
}
\arguments{
\item{data}{A dataset in a Pirat compliant format}
}
\value{
A plot
}
\description{
Make the missingness mechanism plot for Pirat
}
\examples{
data(subbouyssie, package = "Pirat")
missmechPiratPlot(subbouyssie)

}
