% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_build_design.R
\name{mod_build_design_ui}
\alias{mod_build_design_ui}
\alias{mod_build_design_server}
\title{mod_build_design_ui and mod_build_design_server}
\usage{
mod_build_design_ui(id)

mod_build_design_server(input, output, session, sampleNames)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{sampleNames}{A vector de names for the samples in the dataset}
}
\description{
A shiny Module.
}
\keyword{internal}
