% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_choose_pipeline.R
\name{mod_choose_pipeline_ui}
\alias{mod_choose_pipeline_ui}
\alias{mod_choose_pipeline_server}
\title{mod_choose_pipeline_ui and mod_choose_pipeline_server}
\usage{
mod_choose_pipeline_ui(id)

mod_choose_pipeline_server(input, output, session, pipeline.def, dataType)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{pipeline.def}{xxx}
}
\description{
A shiny Module.
}
\keyword{internal}
