% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipelineConvert_Convert.R
\name{mod_convert}
\alias{mod_convert}
\alias{PipelineConvert_Convert_conf}
\alias{PipelineConvert_Convert_ui}
\alias{PipelineConvert_Convert_server}
\title{xxx}
\usage{
PipelineConvert_Convert_conf()

PipelineConvert_Convert_ui(id)

PipelineConvert_Convert_server(
  id,
  dataIn = reactive({
     NULL
 }),
  steps.enabled = reactive({
     NULL
 }),
  remoteReset = reactive({
     NULL
 }),
  steps.status = reactive({
     NULL
 }),
  current.pos = reactive({
     1
 }),
  btnEvents = reactive({
     NULL
 })
)
}
\arguments{
\item{id}{xxx}

\item{dataIn}{xxx}

\item{steps.enabled}{xxx}

\item{remoteReset}{xxx}
}
\value{
NA

NA
}
\description{
xxx

A shiny Module.
}
\examples{
if (interactive()){
data("Exp1_R25_prot", package = "DaparToolshedData")
path <- system.file('workflow/PipelineConvert', package = 'Prostar2')
shiny::runApp(workflowApp("PipelineConvert_Convert", path, dataIn = Exp1_R25_prot))
}

}
\author{
Samuel Wieczorek, Manon Gaudin
}
\keyword{internal}
