% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_convert_ms_file.R
\name{mod_convert_ms_file_ui}
\alias{mod_convert_ms_file_ui}
\alias{mod_convert_ms_file_server}
\title{mod_convert_ms_file_ui and mod_convert_ms_file_server}
\usage{
mod_convert_ms_file_ui(id)

mod_convert_ms_file_server(input, output, session, pipeline.def)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
