% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_import_file_from.R
\name{mod_import_file_from_ui}
\alias{mod_import_file_from_ui}
\alias{mod_import_file_from_server}
\title{mod_import_file_from_ui and mod_import_file_from_server}
\usage{
mod_import_file_from_ui(id)

mod_import_file_from_server(input, output, session, reset = FALSE)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
