% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_infos_dataset.R
\name{mod_infos_dataset_ui}
\alias{mod_infos_dataset_ui}
\alias{mod_infos_dataset_server}
\title{mod_infos_dataset_ui and mod_infos_dataset_server}
\usage{
mod_infos_dataset_ui(id)

mod_infos_dataset_server(input, output, session, obj = NULL)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{obj}{xxxx}
}
\description{
A shiny Module.
}
\keyword{internal}
