% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_inputGroup.R
\name{mod_inputGroup}
\alias{mod_inputGroup}
\alias{mod_inputGroup_ui}
\alias{mod_inputGroup_server}
\title{Help popover windows}
\usage{
mod_inputGroup_ui(id)

mod_inputGroup_server(
  id,
  df = reactive({
     NULL
 }),
  quantCols = reactive({
     NULL
 }),
  remoteReset = reactive({
     0
 }),
  is.enabled = reactive({
     TRUE
 })
)

mod_inputGroup()
}
\arguments{
\item{id}{A \code{character(1)} xxx}

\item{df}{xxx}

\item{quantCols}{A vector of}

\item{remoteReset}{A \code{logical(1)} which acts as a remote command to reset
the module to its default values. Default is FALSE.}

\item{is.enabled}{xxx}
}
\value{
A shiny app
}
\description{
A shiny Module.
}
\examples{
if (interactive()){
shiny::runApp(mod_inputGroup())
}

}
