% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Metacell_Filtering.R
\name{mod_Metacell_Filtering}
\alias{mod_Metacell_Filtering}
\alias{mod_Metacell_Filtering_ui}
\alias{mod_Metacell_Filtering_server}
\title{Filtering Shiny module}
\usage{
mod_Metacell_Filtering_ui(id)

mod_Metacell_Filtering_server(
  id,
  obj = reactive({
     NULL
 }),
  i = reactive({
     1
 }),
  remoteReset = reactive({
     0
 }),
  is.enabled = reactive({
     TRUE
 })
)

mod_Metacell_Filtering(
  obj = NULL,
  i = 1,
  remoteReset = reactive({
     0
 }),
  is.enabled = TRUE
)
}
\arguments{
\item{id}{xxx}

\item{obj}{An instance of the class \code{QFeatures}}

\item{remoteReset}{A \code{ìnteger(1)} xxxx}

\item{is.enabled}{A \code{logical(1)} that indicates whether the module is
enabled or disabled. This is a remote command.}

\item{keep_vs_remove}{A character(1) indicating whether to keep or delete
items. Default value is "delete"}

\item{operator}{xxx}
}
\value{
As for all modules used with \code{MagellanNTK}, the return value is a
\code{list()} of two items:
\itemize{
\item trigger : xxx
\item value: In this case, it contains a list() of three slots:
\itemize{
\item ll.var: a list() of instances of the class \code{Filtering},
\item ll.query: a list of \code{character()} which describe the queries in natural
language,
\item ll.widgets.value: a list of the values of widgets.
}
}
}
\description{
This function is a shiny module to create a list of queries (instances of
the class \code{Filtering} to filter the quantitative metadata of an instance
of the class \code{SummarizedExperiment}).
This function is written with specifications of the package \code{MagellanNTK} so
as to be easily integrated into workflfow compliant with \code{MagellanNTK}.
}
\examples{
\dontrun{
library(Prostar2)
library(shinyBS)
data(Exp1_R25_prot, package = 'DaparToolshedData')
shiny::runApp(mod_Metacell_Filtering(Exp1_R25_prot, 1))
}

}
