% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_msnset_explorer.R
\name{mod_msnset_explorer_ui}
\alias{mod_msnset_explorer_ui}
\alias{mod_msnset_explorer_server}
\title{mod_msnset_explorer_ui and mod_msnset_explorer_server}
\usage{
mod_msnset_explorer_ui(id)

mod_msnset_explorer_server(input, output, session, obj = NULL)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\author{
Samuel Wieczorek
show fData of the MSnset object in a table

Samuel Wieczorek
}
\keyword{MSnset}
\keyword{internal}
\keyword{object}
\keyword{of}
\keyword{pData}
\keyword{show}
\keyword{the}
