% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_navigation.R
\name{mod_navigation_ui}
\alias{mod_navigation_ui}
\alias{mod_navigation_server}
\title{mod_navigation_ui and mod_navigation_server}
\usage{
mod_navigation_ui(id)

mod_navigation_server(input, output, session, pages)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{isDone}{xxxxx}

\item{screens}{xxxxx}

\item{rstFunc}{xxxxx}

\item{iconType}{xxxxxx}
}
\description{
A shiny Module.
}
\keyword{internal}
