% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Pept_Imputation.R
\name{mod_Pept_Imputation}
\alias{mod_Pept_Imputation}
\alias{mod_Pept_Imputation_ui}
\alias{mod_Pept_Imputation_server}
\title{Filtering Shiny module}
\usage{
mod_Pept_Imputation_ui(id)

mod_Pept_Imputation_server(
  id,
  dataIn = reactive({
     NULL
 }),
  i = reactive({
     NULL
 }),
  remoteReset = reactive({
     0
 }),
  is.enabled = reactive({
     TRUE
 })
)

mod_Pept_Imputation(dataIn, i)
}
\arguments{
\item{id}{xxx}

\item{dataIn}{An instance of the class \code{QFeatures}}

\item{i}{xx}

\item{remoteReset}{A \code{integer(1)} xxxx}

\item{is.enabled}{A \code{logical(1)} that indicates whether the module is
enabled or disabled. This is a remote command.}
}
\value{
As for all modules used with \code{MagellanNTK}, the return value is a
\code{list()} of two items:
\itemize{
\item trigger : xxx
\item value: In this case, it contains a list() of three slots:
\itemize{
\item ll.var: a list() of instances of the class \code{Filtering},
\item ll.query: a list of \code{character()} which describe the queries in natural
language,
\item ll.widgets.value: a list of the values of widgets.
}
}
}
\description{
This function is a shiny module to xxx
This function is written with specifications of the package \code{MagellanNTK} so
as to be easily integrated into workflfow compliant with \code{MagellanNTK}.
}
\examples{
\dontrun{
data(Exp1_R25_pept, package = "DaparToolshedData")
obj <- Exp1_R25_pept[seq_len(100)]
shiny::runApp(mod_Pept_Imputation(obj, 1))
}

}
