% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots_intensity.R
\name{mod_plots_intensity_ui}
\alias{mod_plots_intensity_ui}
\alias{mod_plots_intensity_server}
\title{mod_plots_intensity_ui and mod_plots_intensity_server}
\usage{
mod_plots_intensity_ui(id)

mod_plots_intensity_server(input, output, session, obj = NULL)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
