% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots_intensity_plots.R
\name{mod_plots_intensity_plots_ui}
\alias{mod_plots_intensity_plots_ui}
\alias{mod_plots_intensity_plots_server}
\title{mod_plots_intensity_plots_ui and mod_plots_intensity_plots_server}
\usage{
mod_plots_intensity_plots_ui(id)

mod_plots_intensity_plots_server(
  input,
  output,
  session,
  dataIn,
  params = NULL,
  reset,
  base_palette
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
