% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots_missing_values.R
\name{mod_plots_missing_values_ui}
\alias{mod_plots_missing_values_ui}
\alias{mod_plots_missing_values_server}
\title{mod_plots_missing_values_ui and mod_plots_missing_values_server}
\usage{
mod_plots_missing_values_ui(id)

mod_plots_missing_values_server(input, output, session, dataIn)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{dataIn}{xxxx}
}
\description{
A shiny Module.
}
\keyword{internal}
