% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Prot_Imputation_POV.R
\name{mod_Prot_Imputation_POV}
\alias{mod_Prot_Imputation_POV}
\alias{mod_Prot_Imputation_POV_ui}
\alias{mod_Prot_Imputation_POV_server}
\title{Filtering Shiny module}
\usage{
mod_Prot_Imputation_POV_ui(id)

mod_Prot_Imputation_POV_server(
  id,
  obj = reactive({
     NULL
 }),
  i = reactive({
     NULL
 }),
  remoteReset = reactive({
     0
 }),
  is.enabled = reactive({
     TRUE
 })
)

mod_Prot_Imputation_POV(obj, i)
}
\arguments{
\item{id}{xxx}

\item{obj}{An instance of the class \code{QFeatures}}

\item{remoteReset}{A \code{integer(1)} xxxx}

\item{is.enabled}{A \code{logical(1)} that indicates whether the module is
enabled or disabled. This is a remote command.}
}
\value{
As for all modules used with \code{MagellanNTK}, the return value is a
\code{list()} of two items:
\itemize{
\item trigger : xxx
\item value: In this case, it contains a list() of three slots:
\itemize{
\item ll.var: a list() of instances of the class \code{Filtering},
\item ll.query: a list of \code{character()} which describe the queries in natural
language,
\item ll.widgets.value: a list of the values of widgets.
}
}
}
\description{
This function is a shiny module to xxx
This function is written with specifications of the package \code{MagellanNTK} so
as to be easily integrated into workflfow compliant with \code{MagellanNTK}.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(Exp1_R25_pept, package="DaparToolshedData")
obj <- Exp1_R25_pept[seq_len(100)]
shiny::runApp(mod_Prot_Imputation_POV(obj, 1))
\dontshow{\}) # examplesIf}
}
