% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Prot_Normalization.R
\name{mod_Prot_Normalization}
\alias{mod_Prot_Normalization}
\alias{mod_Prot_Normalization_ui}
\alias{mod_Prot_Normalization_server}
\title{Filtering Shiny module}
\usage{
mod_Prot_Normalization_ui(id)

mod_Prot_Normalization_server(
  id,
  dataIn = reactive({
     NULL
 }),
  i = reactive({
     NULL
 }),
  remoteReset = reactive({
     0
 }),
  is.enabled = reactive({
     TRUE
 })
)

mod_Prot_Normalization(dataIn, i)
}
\arguments{
\item{id}{xxx}

\item{dataIn}{An instance of the class \code{QFeatures}}

\item{i}{xxx}

\item{remoteReset}{A \code{logical(1)} which acts as a remote command to reset
the module to its default values. Default is FALSE.}

\item{is.enabled}{A \code{logical(1)} that indicates whether the module is
enabled or disabled. This is a remote command.}
}
\value{
As for all modules used with \code{MagellanNTK}, the return value is a
\code{list()} of two items:
\itemize{
\item trigger : xxx
\item value: In this case, it contains a list() of three slots:
\itemize{
\item ll.var: a list() of instances of the class \code{Filtering},
\item ll.query: a list of \code{character()} which describe the queries in natural
language,
\item ll.widgets.value: a list of the values of widgets.
}
}
}
\description{
This function is a shiny module to xxx
This function is written with specifications of the package \code{MagellanNTK} so
as to be easily integrated into workflfow compliant with \code{MagellanNTK}.
}
\examples{
if (interactive()){
data(Exp1_R25_prot, package = "DaparToolshedData")
obj <- Exp1_R25_prot[seq_len(100)]
mod_Prot_Normalization(obj, 1)
}

}
