% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_set_pval_threshold.R
\name{mod_set_pval_threshold}
\alias{mod_set_pval_threshold}
\alias{mod_set_pval_threshold_ui}
\alias{mod_set_pval_threshold_server}
\title{Module set pval threshold}
\usage{
mod_set_pval_threshold_ui(id)

mod_set_pval_threshold_server(
  id,
  pval_init = reactive({
     1
 }),
  remoteReset = reactive({
     0
 }),
  is.enabled = reactive({
     TRUE
 })
)

mod_set_pval_threshold(pval_init = 1)
}
\arguments{
\item{id}{A \code{charcater()} for the shiny module id}

\item{pval_init}{xxx}

\item{remoteReset}{A \code{logical(1)} which acts as a remote command to reset
the module to its default values. Default is FALSE.}

\item{is.enabled}{xxx}
}
\value{
NA
}
\description{
A shiny Module.
}
\examples{
if (interactive()){
library(Prostar2)
shiny::runApp(mod_set_pval_threshold())
}

}
