% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tracker.R
\name{mod_tracker}
\alias{mod_tracker}
\alias{mod_tracker_ui}
\alias{mod_tracker_server}
\title{Tracking of entities within plots}
\usage{
mod_tracker_ui(id)

mod_tracker_server(
  id,
  object,
  remoteReset = reactive({
     NULL
 }),
  is.enabled = reactive({
     TRUE
 })
)

mod_tracker(obj)
}
\arguments{
\item{id}{xxx}

\item{object}{A instance of the class \code{SummarizedExperiment}}
}
\value{
NA

A \code{list()} of integers
}
\description{
This shiny module offers a UI to select a subset of a dataset and
superimpose quantitative values of this selection on the complete plot
Three modes of selection are implemented:
\itemize{
\item 'Protein list': xxx,
\item 'Random': xxx,
\item 'Specific column': xxx
}
}
