% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_var_dist_plot.R
\name{mod_var_dist_plot_ui}
\alias{mod_var_dist_plot_ui}
\alias{mod_var_dist_plot_server}
\title{mod_var_dist_plot_ui and mod_var_dist_plot_server}
\usage{
mod_var_dist_plot_ui(id)

mod_var_dist_plot_server(input, output, session, obj)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
