% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_funcs.R
\name{normalizationProt}
\alias{normalizationProt}
\title{Protein normalization}
\usage{
normalizationProt(
  data,
  method,
  history = NULL,
  quantile = NULL,
  type = NULL,
  scaling = NULL,
  subset.norm = NULL,
  span = NULL
)
}
\arguments{
\item{data}{A \code{QFeatures}}

\item{method}{A \code{character(1)}, normalization method to be used. Must choose
between : \code{GlobalQuantileAlignment}, \code{SumByColumns}, \code{QuantileCentering},
\code{MeanCentering}, \code{LOESS} and \code{vsn}}

\item{history}{The history}

\item{quantile}{A \code{numeric(1)}, quantile of the intensity distribution that
is used as reference. Must be a float between 0 and 1}

\item{type}{A \code{character(1)}, indicates whether the method is applied to the
entire dataset at once (\code{overall}) or whether each condition is normalized
independently (\verb{within conditions})}

\item{scaling}{A \code{logical(1)}, whether variance reduction is performed}

\item{subset.norm}{Selection of the proteins to which normalization will
be applied}

\item{span}{A \code{numeric(1)}, proportion of the other analytes considered to
perform the regression. Must be a float between 0 and 1}
}
\value{
A \code{list} containing the normalized dataset and the history
}
\description{
Perform normalization for proteins
}
\examples{
data(subR25prot, package = "DaparToolshed")
normalizationProt(subR25prot,
                  method = "GlobalQuantileAlignment")

}
