% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{not_a_numeric}
\alias{not_a_numeric}
\title{Numeric test}
\usage{
not_a_numeric(input)
}
\arguments{
\item{input}{The object to test}
}
\value{
\code{NULL} if numeric, a \code{character} if not
}
\description{
Test whether an object is numeric or not
}
\examples{
not_a_numeric(1)
not_a_numeric("A")

}
