% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_open_dataset.R
\name{open_dataset}
\alias{open_dataset}
\alias{open_dataset_ui}
\alias{open_dataset_server}
\title{open_dataset_ui and open_dataset_server}
\usage{
open_dataset_ui(id)

open_dataset_server(
  id,
  remoteReset = reactive({
     0
 }),
  is.enabled = reactive({
     TRUE
 })
)
}
\arguments{
\item{id}{xxx}
}
\description{
A shiny Module.
}
\examples{
if (interactive()){
ui <- fluidPage(
tagList(
  open_dataset_ui("qf_file"),
  textOutput('res')
)
)

server <- function(input, output, session) {
  rv <- reactiveValues(
    obj = NULL
  )
  rv$obj <- open_dataset_server("qf_file")
  
  output$res <- renderText({
    rv$obj()
    paste0('Names of the datasets: ', names(rv$obj()))
  })
}

shinyApp(ui, server)
}

}
\keyword{internal}
