% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDA_funcs.R
\name{pairwiseComparisonProt}
\alias{pairwiseComparisonProt}
\title{PipelineProtein pairwise comparison sub-step}
\usage{
pairwiseComparisonProt(history, comparison, dt, pushed)
}
\arguments{
\item{history}{The history}

\item{comparison}{A \code{character(1)}, the name of the comparison}

\item{dt}{A \code{data.frame}, the pushed p-values dt}

\item{pushed}{A \code{list} with pushed p-values in the differents comparisons}
}
\value{
A \code{list} containing the history
}
\description{
Do what has to be done at the end of the pairwise comparison
sub-step of the PipelineProtein
}
\examples{
NULL

}
