% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkgs_require}
\alias{pkgs_require}
\title{Loads packages}
\usage{
pkgs_require(ll.deps)
}
\arguments{
\item{ll.deps}{A \code{character()} vector which contains packages names}
}
\value{
NA
}
\description{
Checks if a package is available to load it
}
\examples{
NULL

}
\author{
Samuel Wieczorek
}
