% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_funcs.R
\name{prepareQFsave}
\alias{prepareQFsave}
\title{End process save dataset}
\usage{
prepareQFsave(
  data,
  i = NULL,
  history,
  namePipeline = "PipelineProtein",
  SEname = NULL
)
}
\arguments{
\item{data}{A \code{QFeatures}}

\item{i}{A \code{numeric(1)}, SE to work on}

\item{history}{A \code{data.frame}, contains the history to add to
the designated SE}

\item{namePipeline}{A \code{character(1)}, the pipeline type}

\item{SEname}{A \code{character(1)}, name to give to the designated SE}
}
\value{
A \code{QFeatures}
}
\description{
Do the necessary steps at the end of a process
}
\examples{
NULL

}
