% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prostar2.R
\name{Prostar2}
\alias{Prostar2}
\title{xxx}
\usage{
Prostar2(
  wf.name = NULL,
  convert.name = "PipelineProtein_Convert",
  usermod = "user",
  verbose = FALSE
)
}
\arguments{
\item{wf.name}{xxx}

\item{convert.name}{xxx}

\item{usermod}{xxx}

\item{verbose}{xxx}
}
\value{
NA
}
\description{
xxxx
}
\examples{
if (interactive()){
library(Prostar2)
library(MagellanNTK)
Prostar2("PipelineProtein")
Prostar2("PipelineProtein_Filtering")
Prostar2("PipelineProtein_Normalization")
Prostar2("PipelineProtein_Imputation")
Prostar2("PipelineProtein_HypothesisTest")
}

}
