% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDA_funcs.R
\name{pushPvalues}
\alias{pushPvalues}
\title{Pushed p-values}
\usage{
pushPvalues(data, ind, command, query, comparison, dt, pushed)
}
\arguments{
\item{data}{A \code{QFeatures}}

\item{ind}{The p-values to push}

\item{command}{A \code{character(1)}, the command}

\item{query}{A \code{character(1)}, the query}

\item{comparison}{A \code{character(1)}, the name of the comparison}

\item{dt}{A \code{data.frame}, the pushed p-values dt}

\item{pushed}{A \code{list} with pushed p-values in the differents comparisons}
}
\value{
A \code{list} containing the data, the pushed p-values dt and the pushed
list
}
\description{
Pusing p-values
}
\examples{
NULL

}
