% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDA_funcs.R
\name{pvalCalibrationProt}
\alias{pvalCalibrationProt}
\title{PipelineProtein p-value calibration sub-step}
\usage{
pvalCalibrationProt(history, calibmet, pi0, h1concent, unifunder)
}
\arguments{
\item{history}{The history}

\item{calibmet}{A \code{character(1)}, the calibration method}

\item{pi0}{A \code{numeric(1)}, the pi0}

\item{h1concent}{The h1 concentration}

\item{unifunder}{The Uniformity underestimation}
}
\value{
A list containing the history
}
\description{
Do what has to be done at the end of the p-value calibration
sub-step of the PipelineProtein
}
\examples{
NULL

}
