% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_funcs.R
\name{readFileConvert}
\alias{readFileConvert}
\title{Read file}
\usage{
readFileConvert(file, sheet = NULL)
}
\arguments{
\item{file}{The object from a fileInput widget}

\item{sheet}{The name of the sheet if the dataset is an excel}
}
\value{
The dataset
}
\description{
Read a file uploaded
}
\examples{
NULL

}
