% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_funcs.R
\name{save_txt_ui}
\alias{save_txt_ui}
\title{Save text UI}
\usage{
save_txt_ui()
}
\value{
The corresponding HTML
}
\description{
The UI for the text and its appearance in each save
}
\examples{
save_txt_ui()

}
