% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ScreenManager.R
\name{ScreenManager}
\alias{ScreenManager}
\title{Add together two numbers}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers

Add together two numbers
}
\examples{
add(1, 1)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ScreenManager$new()}}
\item \href{#method-CheckConfig}{\code{ScreenManager$CheckConfig()}}
\item \href{#method-Additional_Initialize_Class}{\code{ScreenManager$Additional_Initialize_Class()}}
\item \href{#method-GetScreens_server}{\code{ScreenManager$GetScreens_server()}}
\item \href{#method-ActionOn_New_DataIn}{\code{ScreenManager$ActionOn_New_DataIn()}}
\item \href{#method-Additional_Server_Funcs}{\code{ScreenManager$Additional_Server_Funcs()}}
\item \href{#method-Set_Skipped}{\code{ScreenManager$Set_Skipped()}}
\item \href{#method-Set_Reseted}{\code{ScreenManager$Set_Reseted()}}
\item \href{#method-ValidateCurrentPos}{\code{ScreenManager$ValidateCurrentPos()}}
\item \href{#method-EncapsulateScreens}{\code{ScreenManager$EncapsulateScreens()}}
\item \href{#method-ActionOn_NewPosition}{\code{ScreenManager$ActionOn_NewPosition()}}
\item \href{#method-GetStringStatus}{\code{ScreenManager$GetStringStatus()}}
\item \href{#method-Timestamp}{\code{ScreenManager$Timestamp()}}
\item \href{#method-Send_Result_to_Caller}{\code{ScreenManager$Send_Result_to_Caller()}}
\item \href{#method-Get_Result}{\code{ScreenManager$Get_Result()}}
\item \href{#method-InitializeDataIn}{\code{ScreenManager$InitializeDataIn()}}
\item \href{#method-dataModal}{\code{ScreenManager$dataModal()}}
\item \href{#method-GetMaxValidated_AllSteps}{\code{ScreenManager$GetMaxValidated_AllSteps()}}
\item \href{#method-SetModalTxt}{\code{ScreenManager$SetModalTxt()}}
\item \href{#method-Change_Current_Pos}{\code{ScreenManager$Change_Current_Pos()}}
\item \href{#method-ToggleState_Screens}{\code{ScreenManager$ToggleState_Screens()}}
\item \href{#method-ToggleState_ResetBtn}{\code{ScreenManager$ToggleState_ResetBtn()}}
\item \href{#method-ResetScreens}{\code{ScreenManager$ResetScreens()}}
\item \href{#method-Initialize_Status_Process}{\code{ScreenManager$Initialize_Status_Process()}}
\item \href{#method-Update_Cursor_Position}{\code{ScreenManager$Update_Cursor_Position()}}
\item \href{#method-GetFirstMandatoryNotValidated}{\code{ScreenManager$GetFirstMandatoryNotValidated()}}
\item \href{#method-BasicReset}{\code{ScreenManager$BasicReset()}}
\item \href{#method-NavPage}{\code{ScreenManager$NavPage()}}
\item \href{#method-Update_State_Screens}{\code{ScreenManager$Update_State_Screens()}}
\item \href{#method-ui}{\code{ScreenManager$ui()}}
\item \href{#method-server}{\code{ScreenManager$server()}}
\item \href{#method-clone}{\code{ScreenManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$new(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckConfig"></a>}}
\if{latex}{\out{\hypertarget{method-CheckConfig}{}}}
\subsection{Method \code{CheckConfig()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$CheckConfig(conf)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Additional_Initialize_Class"></a>}}
\if{latex}{\out{\hypertarget{method-Additional_Initialize_Class}{}}}
\subsection{Method \code{Additional_Initialize_Class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$Additional_Initialize_Class()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetScreens_server"></a>}}
\if{latex}{\out{\hypertarget{method-GetScreens_server}{}}}
\subsection{Method \code{GetScreens_server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$GetScreens_server(input, output)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActionOn_New_DataIn"></a>}}
\if{latex}{\out{\hypertarget{method-ActionOn_New_DataIn}{}}}
\subsection{Method \code{ActionOn_New_DataIn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$ActionOn_New_DataIn()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Additional_Server_Funcs"></a>}}
\if{latex}{\out{\hypertarget{method-Additional_Server_Funcs}{}}}
\subsection{Method \code{Additional_Server_Funcs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$Additional_Server_Funcs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Set_Skipped"></a>}}
\if{latex}{\out{\hypertarget{method-Set_Skipped}{}}}
\subsection{Method \code{Set_Skipped()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$Set_Skipped()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Set_Reseted"></a>}}
\if{latex}{\out{\hypertarget{method-Set_Reseted}{}}}
\subsection{Method \code{Set_Reseted()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$Set_Reseted()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidateCurrentPos"></a>}}
\if{latex}{\out{\hypertarget{method-ValidateCurrentPos}{}}}
\subsection{Method \code{ValidateCurrentPos()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$ValidateCurrentPos()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EncapsulateScreens"></a>}}
\if{latex}{\out{\hypertarget{method-EncapsulateScreens}{}}}
\subsection{Method \code{EncapsulateScreens()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$EncapsulateScreens()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActionOn_NewPosition"></a>}}
\if{latex}{\out{\hypertarget{method-ActionOn_NewPosition}{}}}
\subsection{Method \code{ActionOn_NewPosition()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$ActionOn_NewPosition()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStringStatus"></a>}}
\if{latex}{\out{\hypertarget{method-GetStringStatus}{}}}
\subsection{Method \code{GetStringStatus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$GetStringStatus(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Timestamp"></a>}}
\if{latex}{\out{\hypertarget{method-Timestamp}{}}}
\subsection{Method \code{Timestamp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$Timestamp()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Send_Result_to_Caller"></a>}}
\if{latex}{\out{\hypertarget{method-Send_Result_to_Caller}{}}}
\subsection{Method \code{Send_Result_to_Caller()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$Send_Result_to_Caller()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get_Result"></a>}}
\if{latex}{\out{\hypertarget{method-Get_Result}{}}}
\subsection{Method \code{Get_Result()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$Get_Result()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InitializeDataIn"></a>}}
\if{latex}{\out{\hypertarget{method-InitializeDataIn}{}}}
\subsection{Method \code{InitializeDataIn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$InitializeDataIn()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataModal"></a>}}
\if{latex}{\out{\hypertarget{method-dataModal}{}}}
\subsection{Method \code{dataModal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$dataModal()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetMaxValidated_AllSteps"></a>}}
\if{latex}{\out{\hypertarget{method-GetMaxValidated_AllSteps}{}}}
\subsection{Method \code{GetMaxValidated_AllSteps()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$GetMaxValidated_AllSteps()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SetModalTxt"></a>}}
\if{latex}{\out{\hypertarget{method-SetModalTxt}{}}}
\subsection{Method \code{SetModalTxt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$SetModalTxt(txt)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Change_Current_Pos"></a>}}
\if{latex}{\out{\hypertarget{method-Change_Current_Pos}{}}}
\subsection{Method \code{Change_Current_Pos()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$Change_Current_Pos(i)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ToggleState_Screens"></a>}}
\if{latex}{\out{\hypertarget{method-ToggleState_Screens}{}}}
\subsection{Method \code{ToggleState_Screens()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$ToggleState_Screens(cond, range)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ToggleState_ResetBtn"></a>}}
\if{latex}{\out{\hypertarget{method-ToggleState_ResetBtn}{}}}
\subsection{Method \code{ToggleState_ResetBtn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$ToggleState_ResetBtn(cond)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResetScreens"></a>}}
\if{latex}{\out{\hypertarget{method-ResetScreens}{}}}
\subsection{Method \code{ResetScreens()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$ResetScreens()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Initialize_Status_Process"></a>}}
\if{latex}{\out{\hypertarget{method-Initialize_Status_Process}{}}}
\subsection{Method \code{Initialize_Status_Process()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$Initialize_Status_Process()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Update_Cursor_Position"></a>}}
\if{latex}{\out{\hypertarget{method-Update_Cursor_Position}{}}}
\subsection{Method \code{Update_Cursor_Position()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$Update_Cursor_Position()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetFirstMandatoryNotValidated"></a>}}
\if{latex}{\out{\hypertarget{method-GetFirstMandatoryNotValidated}{}}}
\subsection{Method \code{GetFirstMandatoryNotValidated()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$GetFirstMandatoryNotValidated(range)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicReset"></a>}}
\if{latex}{\out{\hypertarget{method-BasicReset}{}}}
\subsection{Method \code{BasicReset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$BasicReset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NavPage"></a>}}
\if{latex}{\out{\hypertarget{method-NavPage}{}}}
\subsection{Method \code{NavPage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$NavPage(direction)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Update_State_Screens"></a>}}
\if{latex}{\out{\hypertarget{method-Update_State_Screens}{}}}
\subsection{Method \code{Update_State_Screens()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$Update_State_Screens()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ui"></a>}}
\if{latex}{\out{\hypertarget{method-ui}{}}}
\subsection{Method \code{ui()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-server"></a>}}
\if{latex}{\out{\hypertarget{method-server}{}}}
\subsection{Method \code{server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$server(dataIn = reactive({
    NULL
}))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
