% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonFunc.R
\name{shinyValue}
\alias{shinyValue}
\title{#' @export
#' @importFrom shinyjs JS
initComplete <- function(){
  
  return (JS(
    "function(settings, json) {",
    "$(this.api().table().header()).css({'background-color': 'darkgrey', 'color': 'black'});",
    "}"))
}}
\usage{
shinyValue(id, num)
}
\description{
#' @export
#' @importFrom shinyjs JS
initComplete <- function(){
  
  return (JS(
    "function(settings, json) {",
    "$(this.api().table().header()).css({'background-color': 'darkgrey', 'color': 'black'});",
    "}"))
}
}
