% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_log_console}
\alias{show_log_console}
\title{Show log console}
\usage{
show_log_console(
  expr,
  id_notif = NULL,
  cond_info = TRUE,
  cond_warning = TRUE,
  cond_error = TRUE,
  color_info = "grey",
  color_warning = "blue",
  color_error = "red",
  new_first = TRUE,
  console_message = TRUE,
  prefix = "Imputation"
)
}
\arguments{
\item{expr}{xxx}

\item{id_notif}{xxx}

\item{cond_info}{xxx}

\item{cond_warning}{xxx}

\item{cond_error}{xxx}

\item{color_info}{xxx}

\item{color_warning}{xxx}

\item{color_error}{xxx}

\item{new_first}{xxx}

\item{console_message}{xxx}

\item{prefix}{xxx}
}
\value{
xxx
}
\description{
Show R logs in a console in shiny
}
\examples{
NULL

}
