% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDA_funcs.R
\name{swapConditions}
\alias{swapConditions}
\title{Swap conditions logFC}
\usage{
swapConditions(logFC, i)
}
\arguments{
\item{logFC}{A \code{data.frame}, contains all logFC}

\item{i}{A \code{numeric(1)}, which conditions to swap}
}
\value{
A \code{list} containing the new contrast name and the logFC
}
\description{
Swap conditions for logFC plot
}
\examples{
NULL

}
