% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDA_funcs.R
\name{updatePushedDT}
\alias{updatePushedDT}
\title{Update pushed DT}
\usage{
updatePushedDT(dt, comparison)
}
\arguments{
\item{dt}{A \code{data.frame}}

\item{comparison}{A \code{character(1)}, the name of the comparison}
}
\value{
A \code{data.frame}
}
\description{
Update pushed DT
}
\examples{
NULL

}
