% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_funcs.R
\name{variableFiltering}
\alias{variableFiltering}
\title{Variable filtering}
\usage{
variableFiltering(data, widgets, history, original_length)
}
\arguments{
\item{data}{A \code{QFeatures}}

\item{widgets}{A \code{list()}, all widget values}

\item{history}{A \code{data.frame}, the history}

\item{original_length}{A \code{numeric(1)}, the length of the dataset prefiltering}
}
\value{
A \code{list} containing the filtered dataset, important values
and the history
}
\description{
Perform variable filtering
}
\examples{
NULL

}
