% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_funcs.R
\name{Variablefiltering_BuildVariableFilter}
\alias{Variablefiltering_BuildVariableFilter}
\title{Create filter}
\usage{
Variablefiltering_BuildVariableFilter(
  value = NULL,
  operator = NULL,
  cname = NULL,
  keep_vs_remove = NULL,
  data = NULL,
  i = NULL
)
}
\arguments{
\item{value}{A \code{character(1)} or \code{numeric(1)} the value to apply}

\item{operator}{A \code{character(1)} of the operator used in the filter}

\item{cname}{A \code{character(1)} of the column name}

\item{keep_vs_remove}{A \code{character(1)} whether the filter is used to keep
or to delete}

\item{data}{A \code{QFeatures} corresponding to the dataset to which filters
are applied}

\item{i}{A \code{numeric(1)} corresponding to the SummarizedExperiment to use}
}
\value{
A \code{NumericVariableFilter} filter to apply on the dataset
}
\description{
Create filter for variable filtering
}
\examples{
data(subR25prot, package = "DaparToolshed")
Variablefiltering_BuildVariableFilter(value = 2,
                             operator = "<",
                             cname = "Unique_peptides",
                             keep_vs_remove = "delete",
                             data = subR25prot,
                             i = 1)

}
