% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrationPlot.R
\name{wrapperCalibrationPlot}
\alias{wrapperCalibrationPlot}
\title{Performs a calibration plot on an \code{SummarizedExperiment} object,
calling the \code{cp4p} package functions.}
\usage{
wrapperCalibrationPlot(vPVal, pi0Method = "pounds")
}
\arguments{
\item{vPVal}{A dataframe that contains quantitative data.}

\item{pi0Method}{A vector of the conditions (one condition per sample).}
}
\value{
A plot
}
\description{
This function is a wrapper to the calibration.plot method of the
\code{cp4p} package for use with \code{SummarizedExperiment} objects.
}
\examples{
if (interactive()){
library(DaparToolshed)
library(SummarizedExperiment)
data(Exp1_R25_prot, package = "DaparToolshedData")
obj <- Exp1_R25_prot
# Simulate imputation of missing values
obj <- NAIsZero(obj, 1)
obj <- NAIsZero(obj, 2)
qData <- as.matrix(SummarizedExperiment::assay(obj[[2]]))
sTab <- colData(obj)
limma <- limmaCompleteTest(qData, sTab)
wrapperCalibrationPlot(limma$P_Value[, 1])
}

}
\author{
Samuel Wieczorek
}
