% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Add2History}
\alias{Add2History}
\title{Get the last validated step before current position.}
\usage{
Add2History(history, step, substep, param.name, value)
}
\arguments{
\item{history}{A \code{data.frame()}}

\item{step}{A \code{character()}}

\item{substep}{A \code{character()}}

\item{param.name}{A \code{character()}}

\item{value}{The value corresponding to the param.name}
}
\value{
A \code{data.frame()}
}
\description{
This function returns the indice of the last validated step
before the current step.
}
\examples{
history <- InitializeHistory()
Add2History(history, "Example step", "First sub-step", "my param", "THE value")

}
