% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Add2History}
\alias{Add2History}
\title{Get the last validated step before current position.}
\usage{
Add2History(history, process, step.name, param.name, value)
}
\arguments{
\item{history}{xxx}

\item{process}{xxx}

\item{step.name}{xxx}

\item{param.name}{xxx}

\item{value}{xxx}
}
\value{
A \code{integer(1)}
}
\description{
This function returns the indice of the last validated step before
the current step.
}
\examples{
NULL
}
