% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_buildDesign.R
\name{build-design}
\alias{build-design}
\alias{mod_buildDesign_ui}
\alias{mod_buildDesign_server}
\alias{mod_buildDesign}
\title{xxxx}
\usage{
mod_buildDesign_ui(id)

mod_buildDesign_server(
  id,
  quantCols,
  remoteReset = reactive({
     0
 }),
  is.enabled = reactive({
     TRUE
 })
)

mod_buildDesign(quantCols)
}
\arguments{
\item{id}{xxx}

\item{quantCols}{xxx}
}
\value{
NA
}
\description{
xxxx
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
shiny::runApp(mod_buildDesign(letters[seq(6)]))

\dontshow{\}) # examplesIf}
}
