% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchToList.R
\name{catchToList}
\alias{catchToList}
\title{Catch to a list}
\usage{
catchToList(expr)
}
\arguments{
\item{expr}{xxx}
}
\value{
A \code{list} of three items: 'value', 'warnings' and 'error'
}
\description{
Catch to a list
}
\examples{
NULL

}
