% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_Convert.R
\name{Convert_wrapper}
\alias{Convert_wrapper}
\alias{convert_dataset_ui}
\alias{convert_dataset_server}
\alias{convert_dataset}
\title{Wrapper to Convert pipeline}
\usage{
convert_dataset_ui(id)

convert_dataset_server(
  id,
  reset = reactive({
     NULL
 }),
  is.enabled = reactive({
     TRUE
 })
)

convert_dataset()
}
\description{
These functions are inspired by the functions run_workflow() in the package
\code{MagellanNTK}. They wrap the entire workflow into a single function
}
\examples{
\dontrun{
library(DaparToolshed)
library(MagellanNTK)
library(Prostar2)
library(spsComps)
shiny::runApp(Prostar2::convert_dataset())
}

}
