% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_funcs.R
\name{dataset-processing}
\alias{dataset-processing}
\alias{keepDatasets}
\alias{keepDatasets,list-method}
\alias{addDatasets}
\alias{addDatasets,list-method}
\title{Datasets processing}
\arguments{
\item{object}{An object of class \code{list}.}

\item{range}{A xxxx}

\item{dataset}{\code{character(1)} providing the base with respect to which
logarithms are computed. Default is log2.}

\item{name}{A \code{character(1)} naming the new array name.}
}
\value{
An processed object of the same class as \code{object}.
}
\description{
This manual page describes manipulation methods using \link{list} objects. In
index or name \code{i} can be specified to define the array (by name of
index) on which to operate.

The following functions are currently available:
\itemize{
\item \code{keepDatasets(object, range)} keep datasets in object which
are in range
\item \code{addDatasets(object, dataset, name)} add the 'dataset' to the
object (of type list)
\item \code{Save(object, file)} stores the object to a .RData file
}
}
\details{
The object must be of type list. Thetwo functions are implemented here for
of the package which uses MagellanNTK
}
