% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_funcs.R
\name{dataset-processing}
\alias{dataset-processing}
\alias{keepDatasets}
\alias{keepDatasets,list-method}
\alias{addDatasets}
\alias{addDatasets,list-method}
\title{Datasets processing}
\usage{
addDatasets(object, dataset, name)

keepDatasets(object, range = seq(length(object)))
}
\arguments{
\item{object}{An instance of type list. Must get TRUE to inherits(object, 'list')}

\item{dataset}{An instance of class \code{SummarizedExperiment}}

\item{name}{the name to associate to the dataset in the final object}

\item{range}{xxx}
}
\value{
An processed object of the same class as \code{object}.
}
\description{
This manual page describes manipulation methods using \link{list} objects. In
index or name \code{i} can be specified to define the array (by name of
index) on which to operate.

The following functions are currently available:
\itemize{
\item \code{keepDatasets(object, range)} keep datasets in object which
are in range
\item \code{addDatasets(object, dataset, name)} add the 'dataset' to the
object (of type list)
\item \code{Save(object, file)} stores the object to a .RData file
}

This function appends a dataset in the list with customization
if necessary

This function deletes the items not included in the
range parameter
}
\details{
The object must be of type list. Thetwo functions are implemented here for
of the package which uses MagellanNTK
}
