% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_designExample.R
\name{design-example}
\alias{design-example}
\alias{mod_designExample_ui}
\alias{mod_designExample_server}
\alias{mod_designExample}
\title{xxxx}
\usage{
mod_designExample_ui(id)

mod_designExample_server(
  id,
  n,
  remoteReset = reactive({
     0
 }),
  is.enabled = reactive({
     TRUE
 })
)

mod_designExample(n = 1)
}
\arguments{
\item{id}{xxx}

\item{n}{xxx}
}
\value{
NA
}
\description{
xxxx
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
shiny::runApp(mod_designExample(2))
\dontshow{\}) # examplesIf}
}
