% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_build_report.R, R/mod_download_dataset.R
\name{build_report}
\alias{build_report}
\alias{download_dataset}
\alias{download_dataset_ui}
\alias{download_dataset_server}
\title{dl}
\usage{
build_report(data, filename = "myDataset")

download_dataset_ui(id)

download_dataset_server(
  id,
  dataIn = reactive({
     NULL
 }),
  extension = c("xlsx", "qf"),
  widget.type = "Link",
  filename = "myDataset",
  excel.style = NULL,
  remoteReset = reactive({
     0
 }),
  is.enabled = reactive({
     TRUE
 })
)

download_dataset(data, filename = "myDataset")
}
\arguments{
\item{id}{internal}

\item{dataIn}{internal}

\item{extension}{Available values are \code{csv} (default), \code{qf} and \code{Excel}.}

\item{widget.type}{Available values are \code{Button} and \code{Link} (default).}

\item{excel.style}{xxx}

\item{name}{internal}
}
\value{
NA
}
\description{
A shiny Module.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(Exp1_R25_prot, package = 'DaparToolshedData')
shiny::runApp(download_dataset(Exp1_R25_prot))
\dontshow{\}) # examplesIf}
}
