% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Extract_Value}
\alias{Extract_Value}
\title{xxxx}
\usage{
Extract_Value(
  value,
  expected_type = c("numeric", "character", "logical", "factor", "integer")
)
}
\arguments{
\item{value}{xx}

\item{expected_type}{xxx}
}
\value{
NA
}
\description{
xxxx
}
\examples{
NULL

}
