% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filtering_example.R
\name{filtering-example}
\alias{filtering-example}
\alias{mod_filtering_example_ui}
\alias{mod_filtering_example_server}
\alias{mod_filtering_example}
\title{Filtering example Shiny module}
\usage{
mod_filtering_example_ui(id)

mod_filtering_example_server(
  id,
  dataIn = reactive({
     NULL
 }),
  indices = NULL,
  operation = "keep",
  title = "myTitle",
  remoteReset = reactive({
     0
 }),
  is.enabled = reactive({
     TRUE
 })
)

mod_filtering_example(
  obj,
  indices = NULL,
  operation = "keep",
  title = "myTitle"
)
}
\arguments{
\item{id}{xxx}

\item{indices}{xxx}

\item{operation}{A character(1) that indicates whether to keep or remove
lines identified by indices. Available values are 'keep' (default)
or 'delete'}

\item{title}{xxx}

\item{obj}{An instance of the class \code{SummarizedExperiment}}
}
\value{
NA
}
\description{
xxxx
}
\examples{
\dontrun{
data(Exp1_R25_prot, package = 'DaparToolshedData')
obj <- Exp1_R25_prot[[1]]
indices <- 1:5
operation <- "delete"
shiny::runApp(mod_filtering_example(obj, indices, operation))
}

}
