% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_open_dataset.R
\name{generic_mod_open_dataset}
\alias{generic_mod_open_dataset}
\alias{open_dataset_ui}
\alias{open_dataset_server}
\title{Change the default functions in \code{MagellanNTK}}
\usage{
open_dataset_ui(id)

open_dataset_server(
  id,
  class = NULL,
  extension = NULL,
  demo_package = NULL,
  remoteReset = reactive({
     NULL
 }),
  is.enabled = reactive({
     TRUE
 })
)
}
\arguments{
\item{id}{xxx}

\item{class}{xxx}

\item{extension}{xxx}

\item{demo_package}{xxx}
}
\value{
NA
}
\description{
This module allows to change
}
\examples{
shiny::runApp(open_dataset(extension = "df"))

}
