% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetHistory}
\alias{GetHistory}
\title{Get the last validated step before current position.}
\usage{
GetHistory(dataIn, x)
}
\arguments{
\item{dataIn}{xxx}

\item{x}{xxxx}
}
\value{
A \code{integer(1)}
}
\description{
This function returns the indice of the last validated step before
the current step.
}
\examples{
NULL
}
