% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetHistory}
\alias{GetHistory}
\title{Get the history of an assay.}
\usage{
GetHistory(dataIn, x)
}
\arguments{
\item{dataIn}{An instance of the \code{SummarizedExperiment} class}

\item{x}{A \code{character} which is the name of an assay in the dataIn}
}
\value{
A \code{data.frame()}
}
\description{
Get the history of an assay.
}
\examples{
NULL

}
