% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{globalvariables}
\alias{globalvariables}
\alias{btn_success_color}
\alias{optionsBtnClass}
\alias{redBtnClass}
\alias{PrevNextBtnClass}
\alias{actionBtnClass}
\alias{listBrewerPalettes}
\alias{btnVentsMasg}
\title{Global variables}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 17.

An object of class \code{character} of length 1.
}
\usage{
btn_success_color

optionsBtnClass

redBtnClass

PrevNextBtnClass

actionBtnClass

listBrewerPalettes

btnVentsMasg
}
\description{
Global variables
}
\keyword{datasets}
