% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_infos_dataset.R
\name{infos_dataset}
\alias{infos_dataset}
\alias{infos_dataset_ui}
\alias{infos_dataset_server}
\title{infos_dataset_ui and infos_dataset_server}
\usage{
infos_dataset_ui(id)

infos_dataset_server(
  id,
  dataIn = reactive({
     NULL
 }),
  remoteReset = reactive({
     0
 }),
  is.enabled = reactive({
     TRUE
 })
)

infos_dataset(obj)
}
\arguments{
\item{id}{shiny id}

\item{dataIn}{An instance of the class \code{QFeatures}.}
}
\value{
A shiny app
}
\description{
A shiny Module.
}
\examples{
if (interactive()){
data(Exp1_R25_prot, package = "DaparToolshedData")
shiny::runApp(infos_dataset(Exp1_R25_prot))
}

}
\keyword{internal}
