% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{InitializeHistory}
\alias{InitializeHistory}
\title{Get the last validated step before current position.}
\usage{
InitializeHistory()
}
\arguments{
\item{widgets.names}{xxx}
}
\value{
A \code{integer(1)}
}
\description{
This function returns the indice of the last validated step before
the current step.
}
\examples{
.names <- c('A_A', 'A_Z', 'B_Q', 'B_F')
InitializeHistory(.names)

}
