% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_metacell.R
\name{metacell-plots}
\alias{metacell-plots}
\alias{mod_ds_metacell_Histos_ui}
\alias{mod_ds_metacell_Histos_server}
\alias{mod_ds_metacell_Histos}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
mod_ds_metacell_Histos_ui(id)

mod_ds_metacell_Histos_server(
  id,
  dataIn = reactive({
     NULL
 }),
  group = reactive({
     NULL
 }),
  pal = reactive({
     NULL
 }),
  pattern = reactive({
     NULL
 }),
  showSelect = reactive({
     TRUE
 }),
  remoteReset = reactive({
     0
 }),
  is.enabled = reactive({
     TRUE
 })
)

mod_ds_metacell_Histos(obj, group, pal, pattern = NULL, showSelect)
}
\arguments{
\item{id}{xxx}

\item{dataIn}{xxx}

\item{group}{xxx}

\item{pal}{xxx}

\item{pattern}{xxx}

\item{showSelect}{xxx}

\item{remoteReset}{A \code{logical(1)} which acts as a remote command to reset
the module to its default values. Default is FALSE.}

\item{is.enabled}{xxx}

\item{obj}{An instance of the class \code{SummarizedExperiment}}
}
\value{
NA
}
\description{
xxxx
}
\examples{
if (interactive()){
data(Exp1_R25_prot, package = "DaparToolshedData")
grp <- DaparToolshed::design_qf(Exp1_R25_prot)$Condition
shiny::runApp(mod_ds_metacell_Histos(Exp1_R25_prot[[1]], group = grp))

# Test with pattern already defined
pat <- c("Missing MEC", "Missing POV")
shiny::runApp(mod_ds_metacell_Histos(Exp1_R25_prot[[1]], pattern = pat, group = grp))
}

}
