% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_metacell_tree.R
\name{metacell-tree}
\alias{metacell-tree}
\alias{mod_metacell_tree_ui}
\alias{mod_metacell_tree_server}
\alias{mod_metacell_tree}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
mod_metacell_tree_ui(id)

mod_metacell_tree_server(
  id,
  dataIn = reactive({
     NULL
 }),
  remoteReset = reactive({
     0
 }),
  is.enabled = reactive({
     TRUE
 })
)

mod_metacell_tree(
  obj = NULL,
  remoteReset = reactive({
     0
 })
)
}
\arguments{
\item{id}{xxx}

\item{remoteReset}{xxx}

\item{obj}{An instance of the class \code{SummarizedExperiment}}
}
\value{
NA
}
\description{
xxxx
}
\examples{
\dontrun{

shiny::runApp(mod_metacell_tree())

data(Exp1_R25_pept, package = 'DaparToolshedData')
shiny::runApp(mod_metacell_tree(Exp1_R25_pept[[1]]))

data(Exp1_R25_prot, package = 'DaparToolshedData')
shiny::runApp(mod_metacell_tree(Exp1_R25_prot[[1]]))
}

}
