% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DetQuantImpValues.R
\name{mod_DetQuantImpValues}
\alias{mod_DetQuantImpValues}
\alias{mod_DetQuantImpValues_ui}
\alias{mod_DetQuantImpValues_server}
\title{popover_for_help_ui and popover_for_help_server}
\usage{
mod_DetQuantImpValues_ui(id)

mod_DetQuantImpValues_server(
  id,
  obj = reactive({
     NULL
 }),
  quant = reactive({
     1
 }),
  factor = reactive({
     1
 }),
  remoteReset = reactive({
     0
 }),
  is.enabled = reactive({
     TRUE
 })
)

mod_DetQuantImpValues(obj)
}
\arguments{
\item{id}{xxx}

\item{obj}{xxx}

\item{quant}{xxx}

\item{factor}{xxx}
}
\description{
A shiny Module.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(ft_na)
shiny::runApp(mod_DetQuantImpValues(ft_na[[1]]))
\dontshow{\}) # examplesIf}
}
