% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots_msnset_explorer.R
\name{mod_plots_msnset_explorer_ui}
\alias{mod_plots_msnset_explorer_ui}
\alias{mod_plots_msnset_explorer_server}
\title{mod_msnset_explorer_ui and mod_msnset_explorer_server}
\usage{
mod_plots_msnset_explorer_ui(id)

mod_plots_msnset_explorer_server(input, output, session, obj = NULL)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\author{
Samuel Wieczorek
}
\keyword{internal}
