% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Pirat.R
\name{mod_Pirat}
\alias{mod_Pirat}
\alias{mod_Pirat_ui}
\alias{mod_Pirat_server}
\title{Pirat shiny app}
\usage{
mod_Pirat_ui(id)

mod_Pirat_server(
  id,
  dataIn = reactive({
     NULL
 }),
  remoteReset = reactive({
     0
 }),
  is.enabled = reactive({
     TRUE
 }),
  verbose = FALSE
)

mod_Pirat(dataIn)
}
\arguments{
\item{id}{The id of the module}

\item{dataIn}{An instance of the class \code{SummarizedExperiment}.}

\item{remoteReset}{A boolean which indicates whether to reset the widgets or not.}

\item{is.enabled}{xxx}

\item{verbose}{A boolean (FALSE as default) which indicates whether to
display more details on the process}
}
\value{
A shiny app

A shiny app

A shiny app

A shiny app
}
\description{
To be customized
}
\examples{
if (interactive()){
data(subbouyssie)

# Builds the instance of `SummarizedExperiment`
obj.se <- pirat2SE(
  subbouyssie$peptides_ab, subbouyssie$adj,
  subbouyssie$mask_prot_diff, subbouyssie$mask_pep_diff
)

# Launch the app
app <- mod_Pirat(obj.se)
shiny::runApp(app)
}

}
