% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_set_pval_threshold.R
\name{mod_set_pval_threshold}
\alias{mod_set_pval_threshold}
\alias{mod_set_pval_threshold_ui}
\alias{mod_set_pval_threshold_server}
\title{Module set pval threshold}
\usage{
mod_set_pval_threshold_ui(id)

mod_set_pval_threshold_server(
  id,
  pval_init = reactive({
     1
 }),
  fdr = reactive({
     0
 }),
  threshold.type = reactive({
     "logpval"
 }),
  remoteReset = reactive({
     0
 }),
  is.enabled = reactive({
     TRUE
 })
)

mod_set_pval_threshold(pval_init = 1, fdr = 0, threshold.type = "logpval")
}
\arguments{
\item{id}{xxx}

\item{pval_init}{xxx}

\item{fdr}{xxx}

\item{options}{xxx}
}
\description{
A shiny Module.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
shiny::runApp(mod_set_pval_threshold())
\dontshow{\}) # examplesIf}
}
