% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tracker.R
\name{mod_tracker}
\alias{mod_tracker}
\alias{mod_tracker_ui}
\alias{mod_tracker_server}
\title{Tracking of entities within plots}
\usage{
mod_tracker_ui(id)

mod_tracker_server(
  id,
  object = reactive({
     NULL
 }),
  remoteReset = reactive({
     0
 }),
  is.enabled = reactive({
     TRUE
 })
)

mod_tracker(dataIn)
}
\arguments{
\item{id}{shiny id}

\item{object}{A instance of the class \code{SummarizedExperiment}}

\item{remoteReset}{xxx}

\item{is.enabled}{xxx}
}
\value{
NA

A \code{list()} of integers
}
\description{
This shiny module offers a UI to select a subset of a dataset and
superimpose quantitative values of this selection on the complete plot
Three modes of selection are implemented:
\itemize{
\item 'Protein list': xxx,
\item 'Random': xxx,
\item 'Specific column': xxx
}
}
\examples{
\dontrun{
data(Exp1_R25_prot, package = "DaparToolshedData")
shiny::runApp(mod_tracker(Exp1_R25_prot))
}

}
