% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_open_dataset.R
\name{open_dataset}
\alias{open_dataset}
\alias{open_dataset_ui}
\alias{open_dataset_server}
\title{open_dataset_ui and open_dataset_server}
\usage{
open_dataset_ui(id)

open_dataset_server(
  id,
  remoteReset = reactive({
     0
 }),
  is.enabled = reactive({
     TRUE
 })
)

open_dataset()
}
\arguments{
\item{id}{xxx}
}
\description{
A shiny Module.
}
\examples{
\dontrun{
shiny::runApp(open_dataset())
}

}
\keyword{internal}
