% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_open_demoDataset.R
\name{open_demo_dataset}
\alias{open_demo_dataset}
\alias{open_demoDataset_ui}
\alias{open_demoDataset_server}
\alias{open_demoDataset}
\title{mod_open_demo_dataset_ui and mod_open_demo_dataset_server}
\usage{
open_demoDataset_ui(id)

open_demoDataset_server(
  id,
  remoteReset = reactive({
     0
 }),
  is.enabled = reactive({
     TRUE
 })
)

open_demoDataset()
}
\arguments{
\item{id}{xxx}
}
\description{
A shiny Module.
}
\examples{
\dontrun{
shiny::runApp(open_demoDataset())
}


}
\keyword{internal}
