% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prostar2.R
\name{Prostar2}
\alias{Prostar2}
\title{xxx}
\usage{
Prostar2(wf.name)
}
\description{
xxxx
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
Prostar2('PipelineProtein')

\dontshow{\}) # examplesIf}
}
