% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prostar2.R
\name{Prostar2}
\alias{Prostar2}
\title{xxx}
\usage{
Prostar2(wf.name = NULL, usermod = "user", verbose = FALSE)
}
\arguments{
\item{wf.name}{xxx}

\item{usermod}{xxx}

\item{verbose}{xxx}
}
\description{
xxxx
}
\examples{
if (interactive()){
library(Prostar2)
Prostar2("PipelineProtein")
Prostar2("PipelineProtein", user = "dev")
Prostar2("PipelinePeptide")
}

}
