% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prostar2.R
\name{Prostar2}
\alias{Prostar2}
\title{xxx}
\usage{
Prostar2(wf.name = NULL, usermod = "user", verbose = FALSE)
}
\description{
xxxx
}
\examples{
\dontrun{
library(Prostar2)
Prostar2('PipelineProtein')
Prostar2('PipelineProtein', user = 'dev')
Prostar2('PipelinePeptide')

Prostar2('PipelineConvert')
}

}
